--- backend/routes/image.js
@@ -1,19 +1,0 @@
-
-const MessageModel = require('../model/message')
-
-const Upload = async (ctx) => {
-    const file = ctx.request.files;
-    console.log(
-        111,
-        file
-    )
-    ctx.body = {
-        code: 0,
-        data: await MessageModel.findAndReplaceUserInfo()
-    }
-};
-
-
-module.exports = {
-    Upload,
-};
