+++ backend/src/config.js
@@ -1,3 +1,9 @@
+const publicIp = require('public-ip');
+const FRONTEND_PORT = process.env.FRONTEND_PORT || 3000;
+
 module.exports = {
-    privateKey: process.env.PRIVATE_KEY || String(Math.random())
+    privateKey: process.env.PRIVATE_KEY || String(Math.random()),
+    frontendOrigin: `http://localhost:${FRONTEND_PORT}`,
+    // frontendOrigin: `http://${await publicIp.v4()}:${FRONTEND_PORT}`,
+    backendOrigin: `http://localhost:9002`,
 }
